tease apart the module a bit better
