+++ app/src/lib/dispatcher/app-store.ts
@@ -77,7 +77,7 @@ import {
 
 import { openShell } from '../open-shell'
 // TODO: move these files under lib
-import { openInExternalEditor } from '../editors/lookup'
+import { launchExternalEditor } from '../editors'
 import { AccountsStore } from './accounts-store'
 import { RepositoriesStore } from './repositories-store'
 import { validatedRepositoryPath } from './validated-repository-path'
@@ -2178,7 +2178,7 @@ export class AppStore {
   public _openInExternalEditor(path: string): Promise<void> {
     const state = this.getState()
     const program = state.selectedExternalEditor
-    return openInExternalEditor(path, program)
+    return launchExternalEditor(path, program)
   }
 
   /** This shouldn't be called directly. See `Dispatcher`. */
