+++ app/src/lib/editors/lookup.ts
@@ -1,6 +1,3 @@
-import { spawn } from 'child_process'
-
-import { fatalError } from '../../lib/fatal-error'
 import { findApp } from './darwin'
 import {
   findSublimeTextExecutable,
@@ -75,26 +72,3 @@ export function getAvailableEditors(): Promise<ReadonlyArray<EditorLookup>> {
   return Promise.resolve([])
 }
 
-async function getPathToEditor(app: string): Promise<string> {
-  const programs = await getAvailableEditors()
-  const match = programs.find(p => p.app === app)
-  if (match) {
-    return match.path
-  } else {
-    return ''
-  }
-}
-
-export async function openInExternalEditor(
-  path: string,
-  app: string
-): Promise<void> {
-  const editorPath = await getPathToEditor(app)
-
-  if (__WIN32__) {
-    spawn(editorPath, [path])
-    return
-  }
-
-  return fatalError('Unsupported OS')
-}
